[feature/SOLID Principles]: refactor the use of solid principle
