+++ assets/js/config.js
@@ -1,5 +1,4 @@
-"use strict";
 // config.js
-const config = {
+export const config = {
     apiUrl: "https://www.googleapis.com/books/v1/volumes?q=",
 };
