+++ assets/js/script.js
@@ -1,40 +1,40 @@
-"use strict";
-let bookList;
-class Book {
-    form;
-    editIndex;
-    cachedFormData;
+import { Utils } from './utils.js';
+export class Book {
+    utils;
+    formhandler;
+    // public editIndex: number | null;
     sortBy;
     sortBtn;
-    constructor() {
-        this.form = document.getElementById('formData');
+    constructor(utils, formhandler) {
+        this.utils = utils;
+        this.formhandler = formhandler;
         this.sortBy = document.getElementById('sortBy');
         this.sortBtn = document.getElementById('sortBtn');
-        bookList = [];
-        this.editIndex = null;
-        this.cachedFormData = {
-            title: '',
-            author: '',
-            isbn: '',
-            publicationDate: '',
-            price: 0,
-            discountPrice: 0,
-            genre: '',
-            bookAge: '',
-        };
-        this.bindEvents();
+    }
+    setFormHandlerInstance(formHandler) {
+        this.formhandler = formHandler;
     }
     // Method to bid all the Data
     bindEvents() {
-        this.form?.addEventListener('submit', this.handleFormSubmit.bind(this));
+        // Fix for reset button
         document
-            .querySelector('[type="reset"]')
-            ?.addEventListener('click', () => this.handleFormReset(true));
+            .getElementById('resetForm')
+            ?.addEventListener('click', () => {
+            this.formhandler.handleFormReset();
+        });
+        document.querySelectorAll('[id^="editBtn-"]').forEach((btn, index) => {
+            console.log(index);
+            btn.addEventListener('click', () => this.editBook(index));
+        });
+        document.querySelectorAll('[id^="deleteBtn-"]').forEach((btn, index) => {
+            console.log(index);
+            btn.addEventListener('click', () => this.deleteBook(index));
+        });
         this.toggleSortBtn();
     }
     toggleSortBtn() {
         const genreFilter = document.getElementById('genreFilter');
-        if (bookList.length <= 1) {
+        if (this.formhandler.bookList.length <= 1) {
             genreFilter.disabled = true;
             this.sortBy.disabled = true;
             this.sortBtn.disabled = true;
@@ -51,262 +51,16 @@ class Book {
             this.sortBtn?.addEventListener('change', this.sortBook.bind(this));
         }
     }
-    // Method to handle form reset
-    handleFormReset(triggerValidation = true) {
-        this.editIndex = null;
-        const isbnInput = document.getElementById('isbn');
-        isbnInput.disabled = false;
-        this.form?.reset();
-        this.updateSubmitButton('Add Book');
-        const errorFields = [
-            'title-error',
-            'author-error',
-            'isbn-error',
-            'publicationDate-error',
-            'price-error',
-            'discountPrice-error',
-            'genre-error',
-        ];
-        errorFields.forEach((field) => Utils.toggleError(field));
-    }
-    // Method to handle form submission
-    handleFormSubmit(event) {
-        event.preventDefault();
-        this.cachedFormData = this.getFormData();
-        if (!this.validateForm())
-            return;
-        const { title, author, isbn, publicationDate, price, discountPrice, genre, } = this.cachedFormData;
-        const bookAge = this.calculateBookAge(publicationDate).trim();
-        if (bookAge) {
-            if (this.editIndex !== null) {
-                bookList[this.editIndex] = {
-                    ...bookList[this.editIndex],
-                    title,
-                    author,
-                    publicationDate,
-                    price,
-                    discountPrice,
-                    genre,
-                    bookAge,
-                };
-                this.editIndex = null;
-                this.handleFormReset();
-                Utils.showModal('Book Edited Successfully!');
-            }
-            else {
-                bookList.push({
-                    title,
-                    author,
-                    isbn,
-                    publicationDate,
-                    price,
-                    discountPrice,
-                    genre,
-                    bookAge,
-                });
-                Utils.showModal('Book Added Successfully');
-            }
-        }
-        this.updateTableData(bookList);
-        this.bindEvents();
-        this.handleFormReset();
-        this.totalBookCount();
-    }
-    // Get form data and return as an object
-    getFormData() {
-        return {
-            title: document.getElementById('title').value.trim(),
-            author: document.getElementById('author').value.trim(),
-            isbn: document.getElementById('isbn').value.trim(),
-            publicationDate: document.getElementById('publicationDate').value.trim(),
-            price: Number(document.getElementById('listPrice').value.trim()),
-            discountPrice: Number(document.getElementById('discountPrice').value.trim()),
-            genre: document.getElementById('genre').value.trim(),
-            bookAge: '',
-        };
-    }
-    // Validate the form inputs using cached data
-    validateForm() {
-        const { title, author, isbn, publicationDate, price, discountPrice, genre, } = this.cachedFormData;
-        let isValid = true;
-        const fields = { title, author, isbn, publicationDate, genre };
-        Object.entries(fields).forEach(([field, value]) => {
-            if (!value) {
-                Utils.toggleError(`${field}-error`, `${field.charAt(0).toUpperCase() + field.slice(1)} is required.`);
-                isValid = false;
-            }
-            else {
-                Utils.toggleError(`${field}-error`);
-            }
-        });
-        if (price <= 0 || isNaN(price)) {
-            Utils.toggleError('price-error', 'Price must be a valid positive number.');
-            isValid = false;
-        }
-        else {
-            Utils.toggleError('price-error');
-        }
-        if (discountPrice < 0 || isNaN(discountPrice)) {
-            Utils.toggleError('discountPrice-error', 'Discount price must be a valid number.');
-            isValid = false;
-        }
-        else {
-            Utils.toggleError('discountPrice-error');
-        }
-        if (discountPrice > price) {
-            Utils.toggleError('discountPrice-error', 'Discount price cannot be higher than price.');
-            isValid = false;
-        }
-        else {
-            Utils.toggleError('discountPrice-error');
-        }
-        if (isNaN(Number(isbn)) || isbn.length !== 13) {
-            Utils.toggleError('isbn-error', 'ISBN must be 13 digits.');
-            isValid = false;
-        }
-        else {
-            Utils.toggleError('isbn-error');
-        }
-        return isValid;
-    }
-    // Method to calculate the age of the book
-    calculateBookAge(publicationDate) {
-        const currentDate = new Date();
-        const pubDate = new Date(publicationDate);
-        const ageInYears = currentDate.getFullYear() - pubDate.getFullYear();
-        const ageInMonths = currentDate.getMonth() - pubDate.getMonth();
-        const ageInDays = currentDate.getDate() - pubDate.getDate();
-        if (currentDate < pubDate) {
-            alert('Wrong publication date!');
-            return '';
-        }
-        let ageText = ageInYears > 0
-            ? `${ageInYears} year(s)`
-            : ageInMonths > 0
-                ? `${ageInMonths} month(s)`
-                : ageInDays > 0
-                    ? `${ageInDays} day(s)`
-                    : `Less than a day old`;
-        return ageText;
-    }
-    // Update submit button text for editing or adding
-    updateSubmitButton(text) {
-        document.querySelector('[type="submit"]').textContent = text;
-    }
-    // Method to update the table data after adding/editing a book
-    updateTableData(books) {
-        const tableBody = document.querySelector('#tableData tbody');
-        tableBody.innerHTML = '';
-        const rows = books.map((book, index) => this.createTableRow(book, index));
-        tableBody.insertAdjacentHTML('beforeend', rows.join(''));
-    }
-    // Create table row for each book
-    createTableRow(book, index) {
-        return `
-          <tr class="border text-center even:bg-gray-200 odd:bg-white">
-              <td class="border">${book.author}</td>
-              <td class="border">${book.title}</td>
-              <td class="border">${book.isbn}</td>
-              <td class="border">${book.publicationDate}</td>
-              <td class="border">${book.genre}</td>
-              <td class="border">${this.discountCalculation(book.price, book.discountPrice)}</td>
-              <td class="border">${book.bookAge}</td>
-              <td class="border">
-                  <button class="border w-full p-1 bg-indigo-500 text-white hover:bg-indigo-700" onclick="fetchBook.editBook(${index})">Edit</button>
-                  <button class="border w-full p-1 bg-red-500 text-white hover:bg-red-700" onclick="fetchBook.deleteBook(${index})">Delete</button>
-              </td>
-          </tr>
-      `;
-    }
-    // Method to show confirmation modal before deletion
-    showDeleteConfirmationModal(index) {
-        const confirmDeleteModal = document.getElementById('confirmDeleteModal');
-        const confirmDeleteMessage = document.getElementById('confirmDeleteMessage');
-        const confirmDeleteBtn = document.getElementById('confirmDeleteBtn');
-        const cancelDeleteBtn = document.getElementById('cancelDeleteBtn');
-        confirmDeleteMessage.textContent = `Are you sure you want to delete the book titled "${bookList[index].title}"?`;
-        confirmDeleteModal.classList.remove('hidden');
-        confirmDeleteBtn.addEventListener('click', () => {
-            this.deleteBookConfirmed(index);
-            confirmDeleteModal.classList.add('hidden');
-        });
-        cancelDeleteBtn.addEventListener('click', () => {
-            confirmDeleteModal.classList.add('hidden');
-        });
-    }
-    // Method to delete the book after confirmation
-    deleteBookConfirmed(index) {
-        bookList.splice(index, 1);
-        this.updateTableData(bookList);
-        this.totalBookCount();
-        Utils.showModal('Book Deleted Successfully');
-    }
-    // Method to delete specific book from the table
-    deleteBook(index) {
-        this.showDeleteConfirmationModal(index);
-    }
-    // Method to edit a specific book
-    editBook(index) {
-        const book = bookList[index];
-        this.cachedFormData = { ...book };
-        this.setForm(book);
-        this.editIndex = index;
-        this.updateSubmitButton('Update Book');
-    }
-    // Method to count books in table
-    totalBookCount() {
-        const displayTotalBook = document.getElementById('total-iteam');
-        displayTotalBook.innerHTML = `Total Books: ${bookList.length}`;
-    }
-    // Method to set form data
-    setForm(data) {
-        const formElements = {
-            title: document.getElementById('title'),
-            author: document.getElementById('author'),
-            isbn: document.getElementById('isbn'),
-            publicationDate: document.getElementById('publicationDate'),
-            listPrice: document.getElementById('listPrice'),
-            discountPrice: document.getElementById('discountPrice'),
-            genre: document.getElementById('genre'),
-        };
-        formElements.title.value = data.title;
-        formElements.author.value = data.author;
-        formElements.isbn.value = data.isbn;
-        formElements.isbn.disabled = true;
-        formElements.publicationDate.value = data.publicationDate;
-        formElements.listPrice.value = data.price.toString();
-        formElements.discountPrice.value = data.discountPrice.toString();
-        const genreSelect = formElements.genre;
-        const genreOptions = Array.from(genreSelect.options).map((option) => option.value.toLowerCase());
-        const bookGenre = data.genre.trim().toLowerCase();
-        if (!genreOptions.includes(bookGenre)) {
-            const newOption = document.createElement('option');
-            newOption.value = data.genre;
-            newOption.textContent = data.genre;
-            genreSelect.appendChild(newOption);
-        }
-        genreSelect.value = data.genre;
-        formElements.title.focus();
-    }
-    // Method to filter books by genre
-    filterGenre(event) {
-        const selectedGenre = event.target.value
-            .trim()
-            .toLowerCase();
-        const filteredBooks = selectedGenre
-            ? bookList.filter((book) => book.genre.toLowerCase() === selectedGenre)
-            : bookList;
-        this.updateTableData(filteredBooks);
-    }
-    // Method to sort books by title
+    // Method to sort books by title or author
     sortBook() {
         const sortBy = this.sortBy.value;
-        const sortBtn = this.sortBtn.value;
+        const sortBtn = this.sortBtn.value; // Here, you should use the state from sortBy
         const sortedBooks = this.sortBooks(sortBy, sortBtn);
         this.updateTableData(sortedBooks);
     }
+    // Sorting logic based on user selection
     sortBooks(sortBy, sortBtn) {
-        const sortedBooks = [...bookList];
+        const sortedBooks = [...this.formhandler.bookList];
         const direction = sortBtn === 'dsc' ? -1 : 1;
         if (sortBy === 'author') {
             sortedBooks.sort((a, b) => {
@@ -324,6 +78,112 @@ class Book {
         }
         return sortedBooks;
     }
+    // Update submit button text for editing or adding
+    updateSubmitButton(text) {
+        document.querySelector('[type="submit"]').textContent = text;
+    }
+    // Method to update the table data after adding/editing a book
+    updateTableData(books) {
+        const tableBody = document.querySelector('#tableData tbody');
+        tableBody.innerHTML = '';
+        const rows = books.map((book, index) => this.createTableRow(book, index));
+        tableBody.insertAdjacentHTML('beforeend', rows.join(''));
+        this.bindEvents();
+    }
+    // Create table row for each book
+    createTableRow(book, index) {
+        return `
+          <tr class="border text-center even:bg-gray-200 odd:bg-white">
+              <td class="border">${book.author}</td>
+              <td class="border">${book.title}</td>
+              <td class="border">${book.isbn}</td>
+              <td class="border">${book.publicationDate}</td>
+              <td class="border">${book.genre}</td>
+              <td class="border">${this.discountCalculation(book.price, book.discountPrice)}</td>
+              <td class="border">${book.bookAge}</td>
+              <td class="border">
+                  <button id="editBtn-${index}" class="border w-full p-1 bg-indigo-500 text-white hover:bg-indigo-700">Edit</button>
+                  <button id="deleteBtn-${index}" class="border w-full p-1 bg-red-500 text-white hover:bg-red-700">Delete</button>
+              </td>
+          </tr>
+      `;
+    }
+    // Method to show confirmation modal before deletion
+    showDeleteConfirmationModal(index) {
+        const confirmDeleteModal = document.getElementById('confirmDeleteModal');
+        const confirmDeleteMessage = document.getElementById('confirmDeleteMessage');
+        const confirmDeleteBtn = document.getElementById('confirmDeleteBtn');
+        const cancelDeleteBtn = document.getElementById('cancelDeleteBtn');
+        confirmDeleteMessage.textContent = `Are you sure you want to delete the book titled "${this.formhandler.bookList[index].title}"?`;
+        confirmDeleteModal.classList.remove('hidden');
+        confirmDeleteBtn.addEventListener('click', () => {
+            this.deleteBookConfirmed(index);
+            confirmDeleteModal.classList.add('hidden');
+        });
+        cancelDeleteBtn.addEventListener('click', () => {
+            confirmDeleteModal.classList.add('hidden');
+        });
+    }
+    // Method to delete the book after confirmation
+    deleteBookConfirmed(index) {
+        this.formhandler.bookList.splice(index, 1);
+        this.updateTableData(this.formhandler.bookList);
+        // this.bindEvents();
+        this.totalBookCount();
+        Utils.showModal('Book Deleted Successfully');
+    }
+    // Method to delete specific book from the table
+    deleteBook(index) {
+        this.showDeleteConfirmationModal(index);
+    }
+    // Method to edit a specific book
+    editBook(index) {
+        const book = this.formhandler.bookList[index];
+        this.formhandler.cachedFormData = { ...book };
+        this.formhandler.setForm(book);
+        this.formhandler.editIndex = index;
+        this.updateSubmitButton('Update Book');
+    }
+    // Method to count books in table
+    totalBookCount() {
+        const displayTotalBook = document.getElementById('total-iteam');
+        displayTotalBook.innerHTML = `Total Books: ${this.formhandler.bookList.length}`;
+    }
+    // Method to filter books by genre
+    filterGenre(event) {
+        const selectedGenre = event.target.value
+            .trim()
+            .toLowerCase();
+        const filteredBooks = selectedGenre
+            ? this.formhandler.bookList.filter((book) => book.genre.toLowerCase() === selectedGenre)
+            : this.formhandler.bookList;
+        this.updateTableData(filteredBooks);
+    }
+    // Method to sort books by title
+    // public sortBook(): void {
+    //   const sortBy = this.sortBy!.value;
+    //   const sortBtn = this.sortBtn!.value;
+    //   const sortedBooks = this.sortBooks(sortBy, sortBtn);
+    //   this.updateTableData(sortedBooks);
+    // }
+    // sortBooks(sortBy: string, sortBtn: string): formData[] {
+    //   const sortedBooks = [...this.formhandler.bookList];
+    //   const direction = sortBtn === 'dsc' ? -1 : 1;
+    //   if (sortBy === 'author') {
+    //     sortedBooks.sort((a, b) => {
+    //       const authorA = a.author ? String(a.author) : '';
+    //       const authorB = b.author ? String(b.author) : '';
+    //       return direction * authorA.localeCompare(authorB);
+    //     });
+    //   } else if (sortBy === 'title') {
+    //     sortedBooks.sort((a, b) => {
+    //       const titleA = a.title ? String(a.title) : '';
+    //       const titleB = b.title ? String(b.title) : '';
+    //       return direction * titleA.localeCompare(titleB);
+    //     });
+    //   }
+    //   return sortedBooks;
+    // }
     // method to calculate discount
     discountCalculation(price, discountedPrice) {
         if (price === undefined && discountedPrice === undefined) {
@@ -358,4 +218,4 @@ class Book {
         }
     }
 }
-const book = new Book();
+// const book = new Book();
