+++ assets/ts/config.ts
@@ -1,4 +1,4 @@
 // config.js
-const config = {
+export const config = {
   apiUrl: "https://www.googleapis.com/books/v1/volumes?q=",
 };
