+++ prettier.config.js
@@ -1,7 +1,7 @@
 module.exports = {
-  semi: true, 
-  singleQuote: true, 
-  trailingComma: 'all', 
+  semi: true,
+  singleQuote: true,
+  trailingComma: 'all',
   tabWidth: 2,
   useTabs: false,
 };
